RANDMAX does not need overwriting. A local variant brings it down to 4 digits.
RANDMAX is an input in the algo, it does not modify the range of the target
only dilutes it quality. Thus I don't need to override it.

The mod with the max range makes it from 0 to max range so I used a while
loop (unoptimal) to regenerate random number till it is from 1000 to 9999.

Used while loop because I mostly observed that occurance of less than 4 digits
was like 1 in 12 loops but actually should avoid the while and do it in 1 calc.