Reverse the algo to lookup based on target instead of guess value
This is because target is limited.
Still has an error - 6411 (target) and 1111 (guess) says 2C instead of 2B
If there is a C, still need to scan for a B, B takes precedence over C

So:
Outer loop is target
If no match skip to next target
If B skip to next target
If C scan and if a B, switch C to B and skip to next target
(is that enough)?