make humanoid spin by borrowing ideas from spintop :)
to eliminate the NAN values of simulation, I reduced the z-velocity
of the spinner from 200 to 50. Interesting that the lower half of
the humanoid spins better than its top. Also fiddled with freebody
and make it stand on the ground.

My humanoid dances (rotates) like a puppet on a string ..